Changes in WineDatabaseDao

+++ app/src/main/java/com/example/mywine/model/Wine.kt
@@ -4,8 +4,12 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "wines_table")
-data class Wine(val name: String, val color: String, val year: Int, val rate: Int = -1) {
-
+data class Wine(
+    val name: String,
+    val color: String,
+    val year: Int,
+    val rate: Int = -1
+) {
     @PrimaryKey(autoGenerate = true)
-    var id: Int = 0
+    var id: Long = 0L
 }
